aggregator-service/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-dotenv
httpx
psycopg2-binary
google-cloud-bigquery==3.17.2
